Re-organize Timer components
Since we now have multiple Timers we need a better way to organize
them and render them. We can do this with a Timer controller of sorts
that renders the CountUp and CountDown timers accordingly.
